+++ src/bot.js
@@ -24,7 +24,7 @@ const helpCommands = {};
 
 const isPermitted = (info) => {
   return info.serverId &&
-    _.isPermitted(info.channelID,
+    _.isPermitted(info.channel.id,
       info.roleIds,
       WHITELISTED_CHANNELS[info.serverId],
       WHITELISTED_ROLES[info.serverId])
@@ -89,40 +89,40 @@ const pokedBy = {
 };
 
 bot.on('message', (message) => {
-  if (message.user.id === bot.user.id) {
+  if (message.author.id === bot.user.id) {
     return;
   }
 
-  logger.info([message.user.name, message.content]);
+  logger.info([message.author.name, message.content]);
 
   const msgInfo = {
-    serverId: evt.d['guild_id'] || null,
-    roleIds: evt.d.member ? evt.d.member.roles : null,
-    message: message,
+    serverId: message.guild.id || null,
+    roleIds: message.member ? message.member.roles.array().map(r=>r.id) : null,
+    message: message.content,
     bot: bot,
-    channelID: channelID,
-    userID: userID
+    channel: message.channel,
+    user: message.author
   };
 
   let wasPoke = false;
 
   unprotectedCommand(msgInfo, _.triggers.poke,
     (info, command, match) => {
-      if(!pokedBy[userID]){
-        pokedBy[userID] = {
+      if(!pokedBy[info.user.id]){
+        pokedBy[info.user.id] = {
           times: 0
         }
       }
-      const messages = [
-        `I'm here, <@${userID}>`,
-        `Yeah yeah, I'm here, <@${userID}>`,
-        `Just cut it out, <@${userID}>! :angry:`,
-        `I'm not replying about this anymore, <@${userID}>`
+      const messages = ![
+        `I'm here, <@${info.user.id}>`,
+        `Yeah yeah, I'm here, <@${info.user.id}>`,
+        `Just cut it out, <@${info.user.id}>! :angry:`,
+        `I'm not replying about this anymore, <@${info.user.id}>`
       ]
-      if(pokedBy[userID].times < messages.length){
-        chat(bot, channelID, messages[pokedBy[userID].times]);
+      if(pokedBy[info.user.id].times < messages.length){
+        chat(bot, info.channel, messages[pokedBy[info.user.id].times]);
       }
-      pokedBy[userID].times++;
+      pokedBy[info.user.id].times++;
 
       wasPoke = true;
     }, null, "Check if I'm around");
@@ -137,9 +137,9 @@ bot.on('message', (message) => {
         if(processedLinksNumber === totalLinksNumber){
           if (issues && issues.length > 0) {
             if (issues.length > 1) {
-              chat(bot, channelID, `Those look like Jira issues:\n ${issues.join('\n')}`);
+              chat(bot, info.channel, `Those look like Jira issues:\n ${issues.join('\n')}`);
             } else {
-              chat(bot, channelID, `That looks like a Jira issue:\n ${issues[0]}`);
+              chat(bot, info.channel, `That looks like a Jira issue:\n ${issues[0]}`);
             }
           }
         }
@@ -166,7 +166,7 @@ bot.on('message', (message) => {
         const param = (commandArgs.length > 1 ? commandArgs[1] : match[5]).trim().toLowerCase() || null;
 
         if (!param || param === "today") {
-          Staffsquared.absencesToday(msgInfo, (absentees) => {
+          Staffsquared.absencesToday(info, (absentees) => {
             const absenteeNames = absentees
                 .sort((p1, p2) => {
                   if(p1['EventTypeId'] > p2['EventTypeId']){
@@ -181,9 +181,9 @@ bot.on('message', (message) => {
                   return ` * (${Staffsquared.EVENT_TYPES[p['EventTypeId']]}) **${p['FirstName']} ${p['LastName']}**`
                 });
             if (absenteeNames.length > 0) {
-              chat(bot, channelID, `<@${msgInfo.userID}>: According to StaffSquared these people are out of office today:\n${absenteeNames.join('\n')}`);
+              chat(bot, info.channel, `<@${info.user.id}>: According to StaffSquared these people are out of office today:\n${absenteeNames.join('\n')}`);
             } else {
-              chat(bot, channelID, `<@${msgInfo.userID}>: According to StaffSquared, nobody is out of office today.`);
+              chat(bot, info.channel, `<@${info.user.id}>: According to StaffSquared, nobody is out of office today.`);
             }
           });
         } else {
@@ -192,12 +192,12 @@ bot.on('message', (message) => {
             return;
           }
 
-          Staffsquared.absencesFuture(msgInfo, (absentees) => {
+          Staffsquared.absencesFuture(info, (absentees) => {
             const absenteeNames = Staffsquared.getNamesInDateRange(absentees, dateRange);
             if (absenteeNames.length > 0) {
-              chat(bot, channelID, `<@${msgInfo.userID}>: According to StaffSquared these people are off ${param}:\n${absenteeNames.join('\n')}`);
+              chat(bot, info.channel, `<@${info.user.id}>: According to StaffSquared these people are off ${param}:\n${absenteeNames.join('\n')}`);
             } else {
-              chat(bot, channelID, `<@${msgInfo.userID}>: According to StaffSquared, nobody is off ${param}.`);
+              chat(bot, info.channel, `<@${info.user.id}>: According to StaffSquared, nobody is off ${param}.`);
             }
           });
         }
@@ -206,7 +206,7 @@ bot.on('message', (message) => {
 
   protectedCommand(msgInfo, _.triggers.release,
     (info, command, match) => {
-      Gitlab.branchList(null, msgInfo, (branchesString) => {
+      Gitlab.branchList(null, info, (branchesString) => {
         const branches = JSON.parse(branchesString);
         const branchNames = branches.filter((b) => {
             return new RegExp(gitlabConfig.releaseBranchPattern).test(b.name);
@@ -214,7 +214,7 @@ bot.on('message', (message) => {
           .sort((a, b) => {
             return a.commit.committed_date < b.commit.committed_date ? 1 : (a.commit.committed_date > b.commit.committed_date ? -1 : 0);
           });
-        chat(msgInfo.bot, msgInfo.channelID, `<@${msgInfo.userID}>, the release branch is probably \`${branchNames[0].name}\``);
+        chat(info.bot, info.channel, `<@${info.user.id}>, the release branch is probably \`${branchNames[0].name}\``);
       });
     }, null, "I'll try to figure out what the release branch is");
 
@@ -224,15 +224,15 @@ bot.on('message', (message) => {
       if (info.message.indexOf(command) !== -1) {
         filter = (info.message.split(command)[1] || "").trim();
       }
-      Gitlab.branchList(filter, msgInfo);
+      Gitlab.branchList(filter, info);
     }, "[|filter]", "I'll list all the branches in git. Optionally pass a filter to \"grep\" by");
 
   protectedCommand(msgInfo, _.triggers.build,
     (info, command, match) => {
       const b = info.message.indexOf(command) !== -1 ? info.message.split(command)[1].trim() : null;
       const branch = match[7] || b || jenkinsConfig.defaultBranch;
-      chat(info.bot, info.channelID, `Sure, <@${info.userID}>. I've asked Jenkins to build from \`${branch}\`.`);
-      Jenkins.requestBuild(branch, msgInfo);
+      chat(info.bot, info.channel, `Sure, <@${info.user.id}>. I've asked Jenkins to build from \`${branch}\`.`);
+      Jenkins.requestBuild(branch, info);
     }, "[|branch]", `I'll ask jenkins to initiate a build (if \`branch\` is not provided then I'll use \`${jenkinsConfig.defaultBranch}\``);
 
   protectedCommand(msgInfo, _.triggers.cancelBuild,
@@ -249,22 +249,22 @@ bot.on('message', (message) => {
       }
       if (type.toLowerCase() === 'queue') {
         Jenkins.cancelQueue(number, () => {
-          chat(info.bot, info.channelID, `I've cancelled queue item ${number}, <@${info.userID}>.`);
+          chat(info.bot, info.channel, `I've cancelled queue item ${number}, <@${info.user.id}>.`);
         });
       } else {
         Jenkins.cancelBuild(number, () => {
-          chat(info.bot, info.channelID, `I've cancelled build number ${number}, <@${info.userID}>.`);
+          chat(info.bot, info.channel, `I've cancelled build number ${number}, <@${info.user.id}>.`);
         });
       }
     }, "[|build|queue] <number>", "I will cancel a build or a queue item. If \`build\` or \`queue\` is not provided, I'll assume it's a build number");
 
   unprotectedCommand(msgInfo, _.triggers.cloudflareStatus,
     (info, command, match) => {
-      CloudflareStatus.getOutages(info.bot, info.channelID);
+      CloudflareStatus.getOutages(info.bot, info.channel);
     }, null, "Check if there's a cloudflare problem");
 
-  unprotectedCommand(msgInfo, _.triggers.help, () => {
-    chat(bot, channelID, `Here's some info, <@${userID}>:
+  unprotectedCommand(msgInfo, _.triggers.help, (info, command, match) => {
+    chat(info.bot, info.channel, `Here's some info, <@${info.user.id}>:
   Commands:
 ${Object.values(helpCommands).join("\n")}
 
@@ -279,11 +279,11 @@ If anyone asks what the release branch is I'll try to find the latest one too!`)
   }, null, "This info");
 
   if(!wasPoke){
-    if(!pokedBy[userID]){
-      pokedBy[userID] = {
+    if(!pokedBy[msgInfo.user.id]){
+      pokedBy[msgInfo.user.id] = {
         times: 0
       }
     }
-    pokedBy[userID].times = 0;
+    pokedBy[msgInfo.user.id].times = 0;
   }
 });