+++ src/utils/discord-chat.js
@@ -1,8 +1,5 @@
 'use strict';
 
-module.exports = (bot, channelID, message) => {
-  bot.sendMessage({
-    to: channelID,
-    message: `:construction_worker: ${message}`
-  });
+module.exports = (bot, channel, message) => {
+  channel.send(`${message}`);
 };